requirements.txt
phonenumbers
streamlit
python-dotenv
langchain
openai
faiss-cpu
